Split backend services into separate blueprints
